requirements.txt
fastapi
uvicorn
psycopg2-binary
networkx
python-dotenv
pydantic